shell: merge to libtrx
