+++ src/tr2/game/game_flow/sequencer.h
@@ -3,6 +3,3 @@
 #include <libtrx/game/game_flow/sequencer.h>
 
 void GF_InitSequencer(void);
-
-GF_COMMAND GF_DoLevelSequence(
-    const GF_LEVEL *start_level, GF_SEQUENCE_CONTEXT seq_ctx);
