+++ src/tr2/meson.build
@@ -111,7 +111,6 @@ sources = [
   'game/room_draw.c',
   'game/savegame/savegame_bson.c',
   'game/savegame/savegame_legacy.c',
-  'game/shell/common.c',
   'game/stats.c',
   'game/ui/dialogs/stats.c',
   'global/enum_map.c',
